wpt_babel/content-security-policy/style-src/style-src-star-allowed.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="style-src *;" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("* should allow any style");
document.addEventListener("securitypolicyviolation", t.unreached_func("Should not trigger a security policy violation"));
  </script>
  <link href="/content-security-policy/style-src/resources/style-src.css" rel="stylesheet" type="text/css"/>
 </head>
 <body>
  <div id="log">
  </div>
  <div id="content">
   Lorem ipsum
  </div>
  <script>
   "use strict";

t.step(function () {
  var contentEl = document.getElementById("content");
  var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
  assert_equals(marginLeftVal, "2px");
  t.done();
});
  </script>
 </body>
</html>
